Add interfaces, no functionality
